API: Added zip export tests, reorganised tests
Extracted an extra method into helper for reuse.
